move health status to components
