+++ src/ui/src/components/llm/LLMKeyManager.tsx
@@ -6,6 +6,7 @@ import * as z from "zod";
 import { formatErrorMessage } from "@/utils";
 import { ipcClient } from "../../services/ipc-client";
 import type { HealthStatusInfo, LLMConfig } from "../../types";
+import { HealthStatus } from "../health-status/health-status";
 import { Button } from "../ui/button";
 import {
   Dialog,
@@ -14,7 +15,6 @@ import {
   DialogTitle,
   DialogTrigger,
 } from "../ui/dialog";
-import { HealthStatus } from "../ui/health-status";
 import { type LLMProvider, LLMProviderForm } from "./LLMProviderForm";
 
 const schema = z.discriminatedUnion("provider", [
